v2/requirements.txt
Pillow>=10.0.0
reportlab>=4.0.0
pytest>=7.0.0
